solved geocoding bug, changed to leaflet geocoder,converted from float to double

+++ src/main/java/com/map/app/containers/RoutePathContainer.java
@@ -42,7 +42,7 @@ public class RoutePathContainer {
 		//making request
 		GHRequest request = new GHRequest(p.getStartlat(), p.getStartlon(), p.getEndlat(), p.getEndlon()).setProfile(profile).putHint(Parameters.CH.DISABLE, true);;
 		PointList pl = new PointList();
-		HashMap<String,Float> map=new HashMap<>();
+		HashMap<String,Double> map=new HashMap<>();
 		ArrayList<String> ins = new ArrayList<>();
 		try {
 			//getting result
@@ -51,9 +51,9 @@ public class RoutePathContainer {
 				throw new RuntimeException(fullRes.getErrors().toString());
 			}
 			ResponsePath res = fullRes.getBest();
-			map.put("Distance in meters", (float)res.getDistance());
+			map.put("Distance in meters", res.getDistance());
 			System.out.println("Distance in meters: " + res.getDistance());
-			map.put("Time in minutes", (float)(res.getTime() / (60.*1000.)));
+			map.put("Time in minutes", (res.getTime() / (60.*1000.)));
 			System.out.println("Time in minutes: " + res.getTime() / (60.*1000.));
 			InstructionList list = res.getInstructions();
 			for (Instruction ele: list) {
